+++ packages/gcp/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v2.0.0rc1] packages/gcp
 ### Releases
 - [PyPI v2.0.0rc1] https://pypi.org/project/mistralai-gcp/2.0.0rc1 - packages/gcp
+
+## 2026-03-09 19:30:40
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.729.0 (2.841.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v2.0.0] packages/gcp
+### Releases
+- [PyPI v2.0.0] https://pypi.org/project/mistralai-gcp/2.0.0 - packages/gcp